+++ src/lib/components/helpers/Help.tsx
@@ -20,6 +20,7 @@ import {
 } from 'react-icons/ai'
 import { RxQuote, RxShare2 } from 'react-icons/rx'
 import useViewerContext from '../hooks/useViewerContext'
+import { LuLayers } from 'react-icons/lu'
 
 const Help = () => {
   const { t } = useTranslation()
@@ -30,17 +31,18 @@ const Help = () => {
   const [indexPosition, setIndexPosition] = useState(0)
   const [yPositions] = useState<string[]>([
     '23px', // Home
-    '59px', // Edit
-    '92px', // Search
-    '127px', // Citation
-    '163px', // TOC
-    '198px', // Share
-    '232px', // Info
-    '267px', // Help
-    '303px', // Theme change
-    '338px', // Download
-    '373px', // Zoom in
-    '407px', // Zoom out
+    '59px', // Groups
+    '92px', // Edit
+    '127px', // Search
+    '163px', // Citation
+    '198px', // TOC
+    '232px', // Share
+    '267px', // Info
+    '303px', // Help
+    '338px', // Theme change
+    '373px', // Download
+    '407px', // Zoom In
+    '440px', // Zoom out
   ])
 
   // css for each type of help step
@@ -58,6 +60,19 @@ const Help = () => {
           icon: <BiHome size={30} />,
           title: t('homeTitle'),
           description: t('homeDescription'),
+          left: true,
+          positionX: '80px',
+          top: true,
+          positionY: '',
+          class: `${tailwindcss['container']} ${tailwindcss['left-step']}`,
+        }
+      : null,
+    // GROUPS
+    editPackage
+      ? {
+          icon: <LuLayers size={30} />,
+          title: t('groupsTitle'),
+          description: t('groupsDescription'),
           left: true,
           positionX: '80px',
           top: true,